requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.20.1
attrs==19.3.0
braceexpand==0.1.5
chardet==3.0.4
colorama==0.4.3
discord==1.0.1
discord.py==1.3.3
extcolors==0.1.3
funcparserlib==0.3.6
humanize==2.4.0
hy==0.18.0
idna==2.9
import-expression==1.1.1
jishaku==1.18.2.188
multidict==4.7.5
Pillow==7.1.2
processor==0.10.0
python-dotenv==0.13.0
pytz==2020.1
qrcode==6.1
rply==0.7.7
six==1.14.0
Twiggy==0.4.7
twiggy-goodies==0.12.0
websockets==8.1
yarl==1.4.2
